requirements.txt
langchain~=0.3.27
langgraph~=0.6.10
crewai~=0.203.1
openai~=1.109.1
streamlit~=1.50.0
opencv-python~=4.12.0.88
pytesseract~=0.3.13
reportlab~=4.4.4
qrcode~=8.2
pillow~=11.3.0
python-dotenv~=1.1.1
pandas~=2.3.3
uv~=0.9.3
numpy~=2.2.6
GitPython~=3.1.45
gitdb~=4.0.12
typing_extensions~=4.15.0
h11~=0.16.0
cryptography~=46.0.2
ipython~=9.6.0
pip~=23.2.1
attrs~=25.4.0
distro~=1.9.0
protobuf~=6.32.1
filelock~=3.20.0
rsa~=4.9.1
pyasn1~=0.6.1
cffi~=2.0.0
grpcio~=1.75.1
parso~=0.8.5
jedi~=0.19.2
colorama~=0.4.6
pytz~=2025.2
rich~=14.2.0
Pygments~=2.19.2
markdown-it-py~=4.0.0
toml~=0.10.2
requests~=2.32.5
PyYAML~=6.0.3
idna~=3.11
multidict~=6.7.0
propcache~=0.4.1
zipp~=3.23.0
sniffio~=1.3.1
tomli~=2.3.0
importlib_metadata~=8.7.0
pyproject_hooks~=1.2.0
build~=1.3.0
smmap~=5.0.2
click~=8.3.0
httpcore~=1.0.9
httpx~=0.28.1
zstandard~=0.25.0
pywin32~=311
jiter~=0.10.0
json5~=0.12.1
mdurl~=0.1.2
charset-normalizer~=3.4.4
overrides~=7.7.0
networkx~=3.5
jsonpickle~=4.1.1
Jinja2~=3.1.6
regex~=2025.9.18
sympy~=1.14.0
mpmath~=1.3.0
decorator~=5.2.1
typer~=0.19.2
shellingham~=1.5.4
altair~=5.5.0
jsonschema~=4.25.1
narwhals~=2.8.0
pyarrow~=21.0.0
packaging~=25.0
referencing~=0.37.0
traitlets~=5.14.3
litellm~=1.74.9
pydantic~=2.12.2
certifi~=2025.10.5
PyJWT~=2.10.1
tomli_w~=1.2.0
pydantic_core~=2.41.4
pydantic-settings~=2.11.0
chromadb~=1.1.1
portalocker~=2.7.0
pyvis~=0.3.2
json_repair~=0.25.2
blinker~=1.9.0
appdirs~=1.4.4
fsspec~=2025.9.0
aiohttp~=3.13.0
yarl~=1.22.0
pyasn1_modules~=0.4.2
cachetools~=6.2.1
urllib3~=2.3.0
MarkupSafe~=3.0.3
websockets~=15.0.1
anyio~=4.11.0
orjson~=3.11.3
openpyxl~=3.1.5
SQLAlchemy~=2.0.44
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
PyPika~=0.48.9
xxhash~=3.6.0
aiosignal~=1.4.0
frozenlist~=1.8.0
aiohappyeyeballs~=2.6.1
backoff~=2.2.1
ipython_pygments_lexers~=1.1.1
stack-data~=0.6.3
prompt_toolkit~=3.0.52
wcwidth~=0.2.14
annotated-types~=0.7.0
tiktoken~=0.12.0
tokenizers~=0.22.1
posthog~=5.4.0
langchain-core~=0.3.79
six~=1.17.0
tzdata~=2025.2
tornado~=6.5.2
uvicorn~=0.37.0
httptools~=0.7.1
watchfiles~=1.1.1
importlib_resources~=6.5.2
pybase64~=1.4.2
tenacity~=9.1.2
bcrypt~=5.0.0
onnxruntime~=1.23.1
kubernetes~=34.1.0
mmh3~=5.2.0
greenlet~=3.2.4
oauthlib~=3.3.1
et_xmlfile~=2.0.0
watchdog~=6.0.0
asttokens~=3.0.0
executing~=2.2.1
langsmith~=0.4.35
langchain-text-splitters~=0.3.11
ormsgpack~=1.11.0
requests-toolbelt~=1.0.0
pure_eval~=0.2.3
pycparser~=2.23
pypdfium2~=4.30.0
websocket-client~=1.9.0
pyreadline3~=3.5.4
instructor~=1.11.3
tqdm~=4.67.1
jsonref~=1.1.0
docstring_parser~=0.17.0
jsonpointer~=3.0.0
jsonschema-specifications~=2025.9.1
durationpy~=0.10
requests-oauthlib~=2.0.0
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=25.9.23
jsonpatch~=1.33
huggingface-hub~=0.35.3
langchain-openai~=0.3.35